backend/requirements.txt
flask==2.3.3
flask-cors==4.0.0
python-dotenv==1.0.0
google-generativeai==0.3.0
chromadb==0.4.15
sentence-transformers==2.2.2
uuid==1.30
